+++ tests/testthat/test-validate_triangle.R
@@ -149,3 +149,75 @@ test_that(".validate_triangle rejects matrix with empty columns", {
     "Invalid reporting triangle structure. Each column must have"
   )
 })
+
+test_that(".validate_triangle errors when only 0s in LHS of NAs", {
+  zero_triangle <- matrix(
+    c(
+      1, 4, 5, 6,
+      0, 2, 5, 4,
+      0, 6, 7, NA,
+      0, 10, NA, NA,
+      0, NA, NA, NA
+    ),
+    nrow = 5,
+    byrow = TRUE
+  )
+  invalid_mat <- matrix(
+    c(
+      0, 0, 1, 2, 3,
+      0, 0, NA, NA, NA
+    ),
+    byrow = TRUE,
+    nrow = 2
+  )
+  valid_mat <- matrix(
+    c(
+      0, 0, 1, 2, 3,
+      1, 0, NA, NA, NA
+    ),
+    byrow = TRUE,
+    nrow = 2
+  )
+  valid_mat2 <- matrix(
+    c(
+      0, 0, 1, 2, 3,
+      0, 1, NA, NA, NA
+    ),
+    byrow = TRUE,
+    nrow = 2
+  )
+
+  could_be_valid_mat <- matrix(
+    c(
+      1, 4, 5, 7, 1,
+      0, 0, 1, 2, 3,
+      0, 0, NA, NA, NA
+    ),
+    byrow = TRUE,
+    nrow = 3
+  )
+
+  expect_error(.validate_triangle(zero_triangle,
+    max_delay = 3,
+    n = 4
+  ))
+  expect_no_error(.validate_triangle(zero_triangle,
+    max_delay = 3,
+    n = 5
+  ))
+  expect_error(.validate_triangle(invalid_mat,
+    max_delay = 3,
+    n = 4
+  ))
+  expect_no_error(.validate_triangle(valid_mat))
+  expect_no_error(.validate_triangle(valid_mat2))
+
+  expect_no_error(.validate_triangle(could_be_valid_mat,
+    max_delay = 4,
+    n = 3
+  ))
+  expect_error(.validate_triangle(could_be_valid_mat,
+    max_delay = 4,
+    n = 2
+  ))
+})